Create TimeframeSelector component
