+++ packages/frontend/src/components/charts/TimeframeSelector.scss
@@ -1,5 +1,5 @@
-@use '../../../styles/mixins.scss';
-@import '../../../styles/variables.scss';
+@use '../../styles/mixins.scss';
+@import '../../styles/variables.scss';
 
 .TimeframeSelector {
   &--MenuActive & {
@@ -16,7 +16,6 @@
     position: absolute;
     z-index: 10;
     width: 100%;
-    // height: 100%;
     height: max-content;
     left: 0;
     backdrop-filter: $blockBackgroundBlur;
